requirements.txt
# Open Brain - Python Dependencies

# Database
psycopg2-binary>=2.9.9
psycopg[binary]>=3.1.18
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# Vector embeddings
requests>=2.31.0

# MCP Server
mcp[cli]>=1.1.0
fastmcp>=0.2.0

# NLP/Entity extraction
spacy>=3.7.4
nltk>=3.8.1

# Analytics
pandas>=2.2.0
numpy>=1.26.4

# Configuration
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Web Dashboard
streamlit>=1.30.0

# Notifications
requests>=2.31.0
